UnitTestDemoG19/src/main/java/uz/pdp/unittestdemog19/service/StudentService.java
package uz.pdp.unittestdemog19.service;


import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import uz.pdp.unittestdemog19.entity.Student;
import uz.pdp.unittestdemog19.exceptions.BadRequestException;
import uz.pdp.unittestdemog19.exceptions.StudentNotFoundException;
import uz.pdp.unittestdemog19.repository.StudentRepository;

import java.util.List;

@RequiredArgsConstructor
@Service
public class StudentService {

    private final StudentRepository studentRepository;

    public List<Student> getAllStudents() {
        return studentRepository.findAll();
    }

    public Student addStudent(Student student) {
        Boolean existsEmail = studentRepository
                .selectExistsEmail(student.getEmail());
        if (existsEmail) {
            throw new BadRequestException(
                    "Email " + student.getEmail() + " taken");
        }

        return studentRepository.save(student);
    }

    public Student addStudentV2(Student student) {
        Boolean existsEmail = studentRepository
                .selectExistsEmail(student.getEmail());
        if (existsEmail) {
            throw new BadRequestException(
                    "Email " + student.getEmail() + " taken");
        }

        studentRepository.save(student);
        return null;
    }

    public void deleteStudent(Long studentId) {
        if(!studentRepository.existsById(studentId)) {
            throw new StudentNotFoundException(
                    "Student with id " + studentId + " does not exists");
        }
        studentRepository.deleteById(studentId);
    }

    public int iSGreat(int a,int b){
        return a<b?a:b;
    }
}
